requirements.txt
gspread==6.0.2
google-cloud-secret-manager==2.18.3
oauth2client==4.1.3
oracledb==2.1.0
